--- apps/core/src/components/pure/workspace-slider-bar/WorkspaceSelector/index.ts
@@ -1,1 +1,0 @@
-export * from './workspace-selector';
