+++ apps/core/src/components/pure/workspace-slider-bar/user-with-workspace-list/styles.ts
