+++ apps/core/src/components/pure/workspace-slider-bar/workspace-card/index.css.ts
