+++ apps/core/src/components/pure/workspace-slider-bar/workspace-card/loading-icon.tsx
