+++ apps/core/src/components/root-app-sidebar/index.tsx
@@ -20,10 +20,17 @@ import {
 } from '@blocksuite/icons';
 import type { Page } from '@blocksuite/store';
 import { useDroppable } from '@dnd-kit/core';
-import { NoSsr } from '@mui/material';
+import { Popover } from '@toeverything/components/popover';
 import { useAtom } from 'jotai';
-import type { ReactElement } from 'react';
-import React, { useCallback, useEffect, useMemo } from 'react';
+import type { HTMLAttributes, ReactElement } from 'react';
+import {
+  forwardRef,
+  Suspense,
+  useCallback,
+  useEffect,
+  useMemo,
+  useState,
+} from 'react';
 
 import { useHistoryAtom } from '../../atoms/history';
 import { useAppSetting } from '../../atoms/settings';
@@ -33,14 +40,14 @@ import { CollectionsList } from '../pure/workspace-slider-bar/collections';
 import { AddCollectionButton } from '../pure/workspace-slider-bar/collections/add-collection-button';
 import { AddFavouriteButton } from '../pure/workspace-slider-bar/favorite/add-favourite-button';
 import FavoriteList from '../pure/workspace-slider-bar/favorite/favorite-list';
-import { WorkspaceSelector } from '../pure/workspace-slider-bar/WorkspaceSelector';
+import { UserWithWorkspaceList } from '../pure/workspace-slider-bar/user-with-workspace-list';
+import { WorkspaceCard } from '../pure/workspace-slider-bar/workspace-card';
 import ImportPage from './import-page';
 
 export type RootAppSidebarProps = {
   isPublicWorkspace: boolean;
   onOpenQuickSearchModal: () => void;
   onOpenSettingModal: () => void;
-  onOpenWorkspaceListModal: () => void;
   currentWorkspace: AllWorkspace;
   openPage: (pageId: string) => void;
   createPage: () => Page;
@@ -52,7 +59,7 @@ export type RootAppSidebarProps = {
   };
 };
 
-const RouteMenuLinkItem = React.forwardRef<
+const RouteMenuLinkItem = forwardRef<
   HTMLDivElement,
   {
     currentPath: string; // todo: pass through useRouter?
@@ -60,7 +67,7 @@ const RouteMenuLinkItem = React.forwardRef<
     icon: ReactElement;
     children?: ReactElement;
     isDraggedOver?: boolean;
-  } & React.HTMLAttributes<HTMLDivElement>
+  } & HTMLAttributes<HTMLDivElement>
 >(({ currentPath, path, icon, children, isDraggedOver, ...props }, ref) => {
   // Force active style when a page is dragged over
   const active = isDraggedOver || currentPath === path;
@@ -94,7 +101,6 @@ export const RootAppSidebar = ({
   currentPath,
   paths,
   onOpenQuickSearchModal,
-  onOpenWorkspaceListModal,
   onOpenSettingModal,
 }: RootAppSidebarProps): ReactElement => {
   const currentWorkspaceId = currentWorkspace.id;
@@ -102,6 +108,7 @@ export const RootAppSidebar = ({
   const { backToAll } = useCollectionManager(currentCollectionsAtom);
   const blockSuiteWorkspace = currentWorkspace.blockSuiteWorkspace;
   const t = useAFFiNEI18N();
+  const [openUserWorkspaceList, setOpenUserWorkspaceList] = useState(false);
   const onClickNewPage = useCallback(async () => {
     const page = createPage();
     await page.waitForLoaded();
@@ -166,12 +173,25 @@ export const RootAppSidebar = ({
         }
       >
         <SidebarContainer>
-          <NoSsr>
-            <WorkspaceSelector
+          <Popover
+            open={openUserWorkspaceList}
+            content={
+              <Suspense>
+                <UserWithWorkspaceList
+                  onEventEnd={useCallback(() => {
+                    setOpenUserWorkspaceList(false);
+                  }, [])}
+                />
+              </Suspense>
+            }
+          >
+            <WorkspaceCard
               currentWorkspace={currentWorkspace}
-              onClick={onOpenWorkspaceListModal}
+              onClick={useCallback(() => {
+                setOpenUserWorkspaceList(true);
+              }, [])}
             />
-          </NoSsr>
+          </Popover>
           <QuickSearchInput
             data-testid="slider-bar-quick-search-button"
             onClick={onOpenQuickSearchModal}
